Change retract speed slightly to avoid problems with scaling.
When the retract speed is eactly the same as some print speed, it can happen
that the slicer does _not_ change the speed after a retract move. Since the
retract move speed gets scaled by the script, that means that the print speed
afterwards is set way too high. Setting the retract speed a tiny bit different
from all the extrusion speeds avoids this problem.